prepare backend for themes

+++ webproject-backend/src/main/java/de/gmasil/webproject/service/dataimport/DataImportService.java
@@ -31,9 +31,11 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.event.ApplicationReadyEvent;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.context.event.EventListener;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
+import org.springframework.util.StopWatch;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
@@ -58,8 +60,9 @@ import de.gmasil.webproject.service.dataimport.ImportData.ImportVideo;
 import de.gmasil.webproject.service.dataimport.ImportData.ImportVideo.ImportComment;
 import de.gmasil.webproject.service.dataimport.ImportData.ImportVideo.ImportFile;
 import de.gmasil.webproject.service.dataimport.ImportData.ImportVideo.ImportRating;
+import de.gmasil.webproject.service.initialize.InitializeFinishedEvent;
 
-@Component
+@Service
 public class DataImportService {
 
     private static final Logger LOG = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
@@ -95,23 +98,35 @@ public class DataImportService {
     @Autowired
     private VideoRatingRepository ratingRepo;
 
-    @Transactional
+    @Autowired
+    private ApplicationEventPublisher eventPublisher;
+
     @EventListener(ApplicationReadyEvent.class)
-    public void createTestData() throws IOException {
+    public void clean() {
+        if (properties.isEnabled() && properties.isClean()) {
+            LOG.info("Cleaning existing data");
+            deleteAllData();
+        }
+        eventPublisher.publishEvent(new CleanFinishedEvent(this));
+    }
+
+    @Transactional
+    @EventListener(InitializeFinishedEvent.class)
+    public void importData() throws IOException {
         if (properties.isEnabled()) {
-            if (properties.isClean()) {
-                LOG.info("Cleaning existing data before import");
-                deleteAllData();
-            }
             File file = new File(properties.getFile());
             if (file.exists()) {
+                StopWatch watch = new StopWatch();
+                watch.start();
                 LOG.info("Importing data from {}", file.getAbsolutePath());
                 importData(file);
-                LOG.info("Data import finished");
+                watch.stop();
+                LOG.info("Data import finished in {}s", watch.getTotalTimeSeconds());
             } else {
                 LOG.warn("Cannot import data, file does not exist: {}", file.getAbsolutePath());
             }
         }
+        eventPublisher.publishEvent(new DataImportFinishedEvent(this));
     }
 
     private void importData(File file) throws IOException {
